llfilesystem: fix clang-tidy bugprone and narrowing warnings
- lldir.cpp: fix ctor member init order, init result=0, use size_type
  for find() result with npos comparison instead of signed -1
- lldir_win32.cpp: init HANDLE/count at declaration, split
  assignment-in-if-condition for both FindFirstFile calls, use
  size_type for find_last_of result
- lldiriterator.cpp: value-initialize boost::smatch
- lldiskcache.h: replace reserved identifier _LLDISKCACHE with
  LL_LLDISKCACHE_H
- lldiskcache.cpp: explicit static_cast<double> on uintmax_t operand
- llfilesystem.cpp: static_cast on tellg/gcount/tellp narrowing to
  S32, reinterpret_cast replacing C-style char* casts, widen
  time_threshold multiplication to std::time_t
- lllfsthread.cpp: init off=0

+++ indra/llfilesystem/lldiskcache.cpp
@@ -206,7 +206,7 @@ const std::string LLDiskCache::getCacheInfo()
 {
     std::ostringstream cache_info;
 
-    double max_in_mb = mMaxSizeBytes / (1024.0 * 1024.0);
+    double max_in_mb = static_cast<double>(mMaxSizeBytes) / (1024.0 * 1024.0);
     double percent_used = ((F32)dirFileSize(sCacheDir) / (F32)mMaxSizeBytes) * 100.0f;
 
     cache_info << std::fixed;
